chore(Topology/{Paracompact + PartitionOfUnity + Algebra/Order/Compact}): restore a more faithful mathported proof, using the "newer" `push_neg` (#5132)
These are just three cases where a mathported proof had to be changed, due to unwanted behaviour with `push_neg/contradiction`.  Since #5082 is supposed to fix some issues with these tactics, some of these workarounds can be removed.

Note: this was not in any way systematic, I simply `grep`ped the line after a `contra`, then `grep`ped for a `not` and then selected 3 likely candidates.  There are potentially more situations like these.