chore: Cleanup condensed abelian group instance (#5198)
I forgot that we already had the fact that sheaves valued in an abelian category form an abelian category, so this PR removes some of the duplication.

Note: The existing file that had this is in `Topology/Category/Sheaves`, and not in `CategoryTheory/Sites`. 
This should be moved at some point.