fix: make `change` tactic use `show` term directly (#3947)
`change` was using the `show` tactic when operating on the main goal, but it was relying on the programmed behavior of `show` rather than the behavior implied by its docstring, which claims it can be used to *select* a goal. This modifies `change` to apply the exact behavior needed to protect against any future changes to the `show` tactic.